pyproject.toml
[project]
name = "youtube-summarizer"
version = "0.1.0"
description = "Backend API for YouTube video transcription and AI summarization"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "aiohttp>=3.13.3",
    "fal-client>=0.13.0",
    "fastapi>=0.129.0",
    "google-genai>=1.63.0",
    "langchain>=1.2.10",
    "langchain-core>=1.2.12",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "opencc>=1.2.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "yt-dlp>=2026.2.4",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "ipykernel>=7.2.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "rich>=14.3.2",
    "ruff>=0.15.1",
]

[project.scripts]
youtube-summarizer = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["youtube_summarizer"]

[tool.ruff]
# High line-length enables manual formatting control via trailing commas:
# - Add trailing comma → forces multi-line formatting
# - Omit trailing comma → keeps inline (up to 320 chars)
line-length = 180

# Match project's minimum Python version
target-version = "py312"

[tool.ruff.lint]
select = [
    # Ruff defaults (pyflakes + essential pycodestyle)
    "E4", "E7", "E9", "F",
    # High-value lints for catching bugs and improving code quality
    "B", "B9", # flake8-bugbear: common bugs and design problems
    "I", # isort: import sorting and organization
    "UP", # pyupgrade: use modern Python syntax
    "SIM", # flake8-simplify: simplify complex code
    "C4", # flake8-comprehensions: better list/dict/set comprehensions
    "ARG", # flake8-unused-arguments: unused function arguments
    "DTZ", # flake8-datetimez: enforce timezone-aware datetimes
    "G", # flake8-logging-format: logging best practices
    "S", # flake8-bandit: security issue detection
    "RUF", # Ruff-specific rules
    "PT", # flake8-pytest-style: pytest best practices
    "PIE", # flake8-pie: unnecessary code patterns
    "T20", # flake8-print: catch print() statements in production code
]

ignore = [
    "B008", # function call in default arg (needed for FastAPI dependencies)
    "S101", # allow assert statements outside tests
    "S105", # hardcoded password detection (too many false positives)
    "S106", # hardcoded password in function args (too many false positives)
]

# Block auto-fix for changes that need careful review
unfixable = [
    "DTZ", # datetime timezone changes can affect business logic
    "S", # security-related fixes should be manually reviewed
]

[tool.ruff.lint.per-file-ignores]
# __init__.py files commonly re-export symbols
"__init__.py" = ["F401", "F403"]

# Test files have different standards
"tests/**/*" = [
    "S101", # asserts are standard in tests
    "ARG", # pytest fixtures often appear unused
    "T20", # print debugging is fine in tests
]
"test_*.py" = ["S101", "ARG", "T20"]

# Notebooks need looser rules
"*.ipynb" = ["T20", "E402"]

[tool.ruff.lint.isort]
# Treat package_name as first-party for import sorting
known-first-party = ["langchain_playground"]

# Combine imports from same module: from x import a, b
combine-as-imports = true

# Deterministic sorting within import groups
force-sort-within-sections = true

[tool.ruff.format]
# Format code examples inside docstrings
docstring-code-format = true
